requirements.txt
pandas
ipython
ipykernel
jupyter
numpy<2.2
matplotlib
plotly
streamlit
pydantic
metalog
scipy
pyyaml
tabulate
numpy-financial
graphviz
statsmodels